Remove gpu test and move tests to test_contrib_amp
